+++ tangy-form-editor/tangy-form-editor.js
@@ -36,7 +36,8 @@ import "tangy-form/input/tangy-qr.js";
 import "tangy-form/input/tangy-gate.js";
 import "tangy-form/input/tangy-consent.js";
 import "tangy-form/input/tangy-signature.js";
-import "tangy-form/input/tangy-prompt-box.js";
+import "tangy-form/input/tangy-audio-playback.js";
+import "tangy-form/input/tangy-audio-recording.js";
 
 /**
  * `tangy-form-editor`